common functions to lib
